Feat: Role and permission

+++ database/migrations/2023_06_28_134023_add_details_to_users_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             $table->dropColumn(['email', 'email_verified_at']);
-            $table->mediumInteger('identity_number')->nullable(false)->unique();
+            $table->bigInteger('identity_number')->nullable(false)->unique();
             $table->integer('age')->nullable(true);
             $table->mediumText('address')->nullable(true);
             $table->boolean('has_blacklisted')->default(false);
